# frontend/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Jewelry APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "http://157.230.33.37:8080"
        }
    ],
    "paths": {
        "/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "getProductDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequestUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/accounts/{id}": {
            "put": {
                "tags": [
                    "Account"
                ],
                "operationId": "updateAccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputValues"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OutputValues"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "operationId": "deleteAccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OutputValues"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/warranties": {
            "post": {
                "tags": [
                    "Warranty"
                ],
                "operationId": "createWarranty",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputValues"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OutputValues"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "insertNewProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imgFiles",
                                    "productRequest"
                                ],
                                "type": "object",
                                "properties": {
                                    "productRequest": {
                                        "$ref": "#/components/schemas/ProductRequest"
                                    },
                                    "imgFiles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "imgFiles",
                                    "productRequest"
                                ],
                                "type": "object",
                                "properties": {
                                    "productRequest": {
                                        "$ref": "#/components/schemas/ProductRequest"
                                    },
                                    "imgFiles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products/all": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "getAllProductsWithSearchSortFilter",
                "parameters": [
                    {
                        "name": "searchKeyword",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "productType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "metalGroup",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "metalTypes",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "gemstoneType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "saleStatus",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "uniqueItems": true,
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ASC",
                                "DESC"
                            ]
                        }
                    },
                    {
                        "name": "pageNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductPriceDto": {
                "type": "object",
                "properties": {
                    "productPriceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMetalCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "totalGemstoneCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "laborCost": {
                        "type": "number",
                        "format": "float"
                    },
                    "markupPercentage": {
                        "maximum": 1,
                        "minimum": 0,
                        "type": "number",
                        "format": "float"
                    },
                    "salePrice": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "ProductRequestUpdate": {
                "required": [
                    "productId",
                    "productName",
                    "productPrice",
                    "productType",
                    "saleStatus",
                    "unitMeasure"
                ],
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "saleStatus": {
                        "type": "string"
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "productPrice": {
                        "$ref": "#/components/schemas/ProductPriceDto"
                    }
                }
            },
...
            "InputValues": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "MANAGER",
                            "STAFF",
                            "CUSTOMER"
                        ]
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "phonenumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "OutputValues": {
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "username": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "MANAGER",
                            "STAFF",
                            "CUSTOMER"
                        ]
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "phonenumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "GemstoneDto": {
                "required": [
                    "gemstoneName"
                ],
                "type": "object",
                "properties": {
                    "gemstoneId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gemstoneName": {
                        "type": "string"
                    },
                    "stoneColor": {
                        "type": "string"
                    },
                    "stoneCut": {
                        "type": "string"
                    },
                    "stoneClarity": {
                        "type": "string"
                    },
                    "materialWeight": {
                        "type": "number",
                        "format": "float"
                    },
                    "materialSize": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "MetalTypeDto": {
                "required": [
                    "materialWeight",
                    "metalTypeName"
                ],
                "type": "object",
                "properties": {
                    "metalTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metalTypeName": {
                        "type": "string"
                    },
                    "metalPurity": {
                        "maximum": 1,
                        "minimum": 0,
                        "type": "number",
                        "format": "float"
                    },
                    "metalGroupName": {
                        "type": "string"
                    },
                    "materialWeight": {
                        "type": "number",
                        "format": "float"
                    },
                    "materialSize": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "ProductRequest": {
                "required": [
                    "metalType",
                    "name",
                    "productPrice",
                    "productType",
                    "saleStatus",
                    "unitMeasure"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 255,
                        "minLength": 5,
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "saleStatus": {
                        "type": "string"
                    },
                    "gemstones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GemstoneDto"
                        }
                    },
                    "metalType": {
                        "$ref": "#/components/schemas/MetalTypeDto"
                    },
                    "productPrice": {
                        "$ref": "#/components/schemas/ProductPriceDto"
                    }
                }
            },
...
            "GemStone": {
                "type": "object",
                "properties": {
                    "gemstoneId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gemstoneName": {
                        "type": "string"
                    },
                    "stoneColor": {
                        "type": "string"
                    },
                    "stoneCut": {
                        "type": "string"
                    },
                    "stoneClarity": {
                        "type": "string"
                    }
                }
            },
            "MetalPriceRate": {
                "type": "object",
                "properties": {
                    "metalPriceRateId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metalTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "profitSell": {
                        "type": "number",
                        "format": "double"
                    },
                    "profitBuy": {
                        "type": "number",
                        "format": "double"
                    },
                    "exchangeRate": {
                        "type": "number",
                        "format": "double"
                    },
                    "internationalPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "metalPriceSpot": {
                        "type": "number",
                        "format": "double"
                    },
                    "sellingPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "buyingPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "effectiveDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "MetalType": {
                "type": "object",
                "properties": {
                    "metalTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metalTypeName": {
                        "type": "string"
                    },
                    "metalPurity": {
                        "type": "number",
                        "format": "double"
                    },
                    "metalGroupName": {
                        "type": "string"
                    },
                    "isAutoUpdatePrice": {
                        "type": "boolean"
                    },
                    "isOnMonitor": {
                        "type": "boolean"
                    },
                    "currentMetalPriceRate": {
                        "$ref": "#/components/schemas/MetalPriceRate"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "unitMeasure": {
                        "type": "string"
                    },
                    "saleStatus": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedBy": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productMaterials": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductMaterial"
                        }
                    },
                    "productPrice": {
                        "$ref": "#/components/schemas/ProductPrice"
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ProductMaterial": {
                "type": "object",
                "properties": {
                    "productMaterialId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "materialId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gemStone": {
                        "$ref": "#/components/schemas/GemStone"
                    },
                    "metalType": {
                        "$ref": "#/components/schemas/MetalType"
                    },
                    "materialWeight": {
                        "type": "number",
                        "format": "float"
                    },
                    "materialSize": {
                        "type": "number",
                        "format": "float"
                    },
                    "metal": {
                        "type": "boolean"
                    }
                }
            },
            "ProductPrice": {
                "type": "object",
                "properties": {
                    "productPriceId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMetalCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "totalGemstoneCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "laborCost": {
                        "type": "number",
                        "format": "double"
                    },
                    "markupPercentage": {
                        "type": "number",
                        "format": "float"
                    },
                    "salePrice": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}